refactor: import keyboard arr from module keyboard.js
