+++ index.html
@@ -8,6 +8,6 @@
     <title>Virtual Keyboard</title>
 </head>
 <body>
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
