+++ keyboard.js
@@ -1,0 +1,445 @@
+const keyboard = [
+  {
+    keyRus: '`',
+    keyEng: ']',
+    code: 'Backquote',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '1',
+    keyEng: '1',
+    code: 'Digit1',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '2',
+    keyEng: '2',
+    code: 'Digit2',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '3',
+    keyEng: '3',
+    code: 'Digit3',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '4',
+    keyEng: '4',
+    code: 'Digit4',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '5',
+    keyEng: '5',
+    code: 'Digit5',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '6',
+    keyEng: '6',
+    code: 'Digit6',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '7',
+    keyEng: '7',
+    code: 'Digit7',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '8',
+    keyEng: '8',
+    code: 'Digit8',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '9',
+    keyEng: '9',
+    code: 'Digit9',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '0',
+    keyEng: '0',
+    code: 'Digit0',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '-',
+    keyEng: '-',
+    code: 'Minus',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '=',
+    keyEng: '=',
+    code: 'Equal',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'Backspace',
+    keyEng: 'Backspace',
+    code: 'Backspace',
+    classes: 'btn btn_func backspace',
+
+  },
+  {
+    keyRus: 'Tab',
+    keyEng: 'Tab',
+    code: 'Tab',
+    classes: 'btn btn_func tab',
+
+  },
+  {
+    keyRus: 'й',
+    keyEng: 'q',
+    code: 'KeyQ',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ц',
+    keyEng: 'w',
+    code: 'KeyW',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'у',
+    keyEng: 'e',
+    code: 'KeyE',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'к',
+    keyEng: 'r',
+    code: 'KeyR',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'е',
+    keyEng: 't',
+    code: 'KeyT',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'н',
+    keyEng: 'y',
+    code: 'KeyY',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'г',
+    keyEng: 'u',
+    code: 'KeyU',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ш',
+    keyEng: 'i',
+    code: 'KeyI',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'щ',
+    keyEng: 'o',
+    code: 'KeyO',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'з',
+    keyEng: 'p',
+    code: 'KeyP',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'х',
+    keyEng: '[',
+    code: 'BracketLeft',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ъ',
+    keyEng: ']',
+    code: 'BracketRight',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ё',
+    keyEng: '\\',
+    code: 'Backslash',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'CapsLock',
+    keyEng: 'CapsLock',
+    code: 'CapsLock',
+    classes: 'btn btn_func caps-lock',
+
+  },
+  {
+    keyRus: 'ф',
+    keyEng: 'a',
+    code: 'KeyA',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ы',
+    keyEng: 's',
+    code: 'KeyS',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'в',
+    keyEng: 'd',
+    code: 'KeyD',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'а',
+    keyEng: 'f',
+    code: 'KeyF',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'п',
+    keyEng: 'g',
+    code: 'KeyG',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'р',
+    keyEng: 'h',
+    code: 'KeyH',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'о',
+    keyEng: 'j',
+    code: 'KeyJ',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'л',
+    keyEng: 'k',
+    code: 'KeyK',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'д',
+    keyEng: 'l',
+    code: 'KeyL',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ж',
+    keyEng: ';',
+    code: 'Semicolon',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'э',
+    keyEng: "'",
+    code: 'Quote',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'Enter',
+    keyEng: 'Enter',
+    code: 'Enter',
+    classes: 'btn btn_func enter',
+
+  },
+  {
+    keyRus: 'Shift',
+    keyEng: 'Shift',
+    code: 'ShiftLeft',
+    classes: 'btn btn_func shift-left',
+
+  },
+  {
+    keyRus: 'я',
+    keyEng: 'z',
+    code: 'KeyZ',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ч',
+    keyEng: 'x',
+    code: 'KeyX',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'с',
+    keyEng: 'c',
+    code: 'KeyC',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'м',
+    keyEng: 'v',
+    code: 'KeyV',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'и',
+    keyEng: 'b',
+    code: 'KeyB',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'т',
+    keyEng: 'n',
+    code: 'KeyN',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ь',
+    keyEng: 'm',
+    code: 'KeyM',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'б',
+    keyEng: ',',
+    code: 'Comma',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: 'ю',
+    keyEng: '.',
+    code: 'Period',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '/',
+    keyEng: '/',
+    code: 'Slash',
+    classes: 'btn',
+
+  },
+  {
+    keyRus: '▲',
+    keyEng: '▲',
+    code: 'ArrowUp',
+    classes: 'btn btn_func',
+
+  },
+  {
+    keyRus: 'Shift',
+    keyEng: 'Shift',
+    code: 'ShiftRight',
+    classes: 'btn btn_func shift-right',
+
+  },
+  {
+    keyRus: 'Control',
+    keyEng: 'Control',
+    code: 'ControlLeft',
+    classes: 'btn btn_func ctrl',
+
+  },
+  {
+    keyRus: 'option',
+    keyEng: 'option',
+    code: 'AltLeft',
+    classes: 'btn btn_func alt',
+
+  },
+  {
+    keyRus: 'command',
+    keyEng: 'command',
+    code: 'MetaLeft',
+    classes: 'btn btn_func meta',
+
+  },
+  {
+    keyRus: 'Space',
+    keyEng: 'Space',
+    code: 'Space',
+    classes: 'btn btn_func space',
+
+  },
+  {
+    keyRus: 'command',
+    keyEng: 'command',
+    code: 'MetaRight',
+    classes: 'btn btn_func meta',
+
+  },
+  {
+    keyRus: 'option',
+    keyEng: 'option',
+    code: 'AltRight',
+    classes: 'btn btn_func option',
+
+  },
+  {
+    keyRus: '◄',
+    keyEng: '◄',
+    code: 'ArrowLeft',
+    classes: 'btn btn_func',
+
+  },
+  {
+    keyRus: '▼',
+    keyEng: '▼',
+    code: 'ArrowDown',
+    classes: 'btn btn_func',
+
+  },
+  {
+    keyRus: '►',
+    keyEng: '►',
+    code: 'ArrowRight',
+    classes: 'btn btn_func',
+
+  },
+];
+
+export default keyboard;
